feat: create the model to contains the cartItem types and remove unnecessary file after refactor
