--- react-client/src/components/Membership/utils/cartTypes.ts
@@ -1,19 +1,0 @@
-export type CartDrawerProps = {
-  isOpen: boolean;
-  onClose: () => void;
-  items?: CartItem[];
-};
-
-export type CartItem = {
-  id: string;
-  name: string;
-  price: string;
-  purchaseLimit?: number;
-  quantity: number;
-  recurrence?: string;
-  acquisitionMessage?: string;
-  type: CartItemType;
-  description: string;
-};
-
-type CartItemType = "product" | "service";
