+++ react-client/src/domain/models/cart/CartModel.ts
@@ -1,0 +1,15 @@
+import { PetCommon } from "../pet/PetModel";
+
+export type AnimalInfo = Pick<PetCommon, "id" | "name">;
+
+export type CartItem = {
+  acquisitionMessage?: string;
+  description: string;
+  id: string;
+  name: string;
+  price: string;
+  purchaseLimit?: number;
+  quantity: number;
+  recurrence?: string;
+  type: string;
+};
